Move interceptors TCK tests using extensions under Full test group.
