+++ impl/src/main/java/org/jboss/cdi/tck/interceptors/tests/full/contract/lifecycleCallback/wrapped/Bird.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.cdi.tck.interceptors.tests.contract.lifecycleCallback.wrapped;
+package org.jboss.cdi.tck.interceptors.tests.full.contract.lifecycleCallback.wrapped;
 
 import static org.testng.Assert.assertNotNull;
